+++ OSHelloWorld/OSHelloWorldApp.swift
@@ -10,13 +10,8 @@ import SwiftUI
 @main
 struct OSHelloWorldApp: App {
     var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-        .defaultSize(width: 100, height: 100)
-
         ImmersiveSpace(id: "ImmersiveSpace") {
-            ImmersiveView()
+            ContentView()
         }
     }
 }
